scsi: ch: avoid false format-overflow warning
gcc-8 notices that printing a 32-bit number into the device name
can overflow the available buffer space:

drivers/scsi/ch.c: In function 'ch_probe':
drivers/scsi/ch.c:937:22: error: '%d' directive writing between 1 and 10 bytes into a region of size 6 [-Werror=format-overflow=]
  sprintf(ch->name,"ch%d",ch->minor);
                      ^~
drivers/scsi/ch.c:937:19: note: directive argument in the range [0, 2147483647]
  sprintf(ch->name,"ch%d",ch->minor);
                   ^~~~~~
drivers/scsi/ch.c:937:2: note: 'sprintf' output between 4 and 13 bytes into a destination of size 8
  sprintf(ch->name,"ch%d",ch->minor);

However, the number comes from an idr_alloc() with an upper bound of '128' so it
will in fact always fit. Using a type cast to a 'u8' type, we can convince gcc
that this code is safe.

+++ drivers/scsi/ch.c
@@ -934,7 +934,7 @@ static int ch_probe(struct device *dev)
 	}
 
 	ch->minor = ret;
-	sprintf(ch->name,"ch%d",ch->minor);
+	sprintf(ch->name, "ch%d", (u8)ch->minor);
 
 	class_dev = device_create(ch_sysfs_class, dev,
 				  MKDEV(SCSI_CHANGER_MAJOR, ch->minor), ch,
